refactor(core): clean architecture with framework-agnostic core
• Move Coordinates from models to core package for domain purity
• Rename adapters.play → adapters.playhttp for clarity
• Add IngestService.handle() in core for future publisher integration
• Update controller flow: parse → validate (core) → handle (core) → map error (adapter)
• Maintain proper dependency direction: controllers.play → adapters.playhttp → core
• Keep error mapping isolated in adapter layer only
• Mirror test packages for clean separation
