+++ src/main/scala/core/Coordinates.scala
@@ -1,3 +1,3 @@
-package models
+package core
 
 case class Coordinates(longitude: Double, latitude: Double)
