+++ src/main/scala/core/IngestService.scala
@@ -1,0 +1,7 @@
+package core
+
+object IngestService {
+  def handle(update: LocationUpdate): Unit = {
+    // No-op for now - will be enhanced when adding publisher/redis integration
+  }
+}
